Rename views.py -> view_helpers.py
