+++ core/views/core_background_tasks.py
@@ -4,7 +4,7 @@ from django.shortcuts import render, redirect
 
 from core import models
 
-from .views import breadcrumb_parser
+from .view_helpers import breadcrumb_parser
 
 logger =logging.getLogger(__name__)
