+++ core/views/external_background_tasks.py
@@ -11,7 +11,7 @@ from core.celery import celery_app
 
 from core.mongo import *
 
-from .views import breadcrumb_parser
+from .view_helpers import breadcrumb_parser
 
 logger = logging.getLogger(__name__)
