+++ core/views/organization.py
@@ -5,7 +5,7 @@ from django.shortcuts import render, redirect
 
 from core import forms, models, tasks
 
-from .views import breadcrumb_parser
+from .view_helpers import breadcrumb_parser
 
 logger = logging.getLogger(__name__)
